refactor proposal operation routes
